added confirm to delete fragment action
git-svn-id: https://source.jasig.org/uPortal/trunk@8436 f5dbab47-78f9-eb45-b975-e544023573eb

+++ webpages/stylesheets/org/jasig/portal/channels/CFragmentManager/html.xsl
@@ -210,13 +210,13 @@ Version $Revision$
                                 </tr>
                                 <tr>
                                     <td>
-                                        <a href="{$baseActionURL}?uPcFM_action=delete&amp;uPcFM_selectedID={$uPcFM_selectedID}">
+                                        <a href="{$baseActionURL}?uPcFM_action=delete&amp;uPcFM_selectedID={$uPcFM_selectedID}" onClick="return confirm('Are you sure you want to delete this fragment?')">
                                             <img src="{$mediaPath}/delete.gif" alt="delete" title="delete" border="0"/>
                                         </a>
                                     </td>
                                     <td width="100%">
                                         <span class="uportal-label">
-                                            <a href="{$baseActionURL}?uPcFM_action=delete&amp;uPcFM_selectedID={$uPcFM_selectedID}" class="uportal-channel-subtitle-reversed">Delete fragment</a>
+                                            <a href="{$baseActionURL}?uPcFM_action=delete&amp;uPcFM_selectedID={$uPcFM_selectedID}" class="uportal-channel-subtitle-reversed"  onClick="return confirm('Are you sure you want to delete this fragment?')">Delete fragment</a>
                                         </span>
                                     </td>
                                 </tr>
